Pequeñas actualizaciones

+++ controllers/ProformaitemController.php
@@ -96,11 +96,18 @@ class ProformaitemController extends Controller
      * @param string $id
      * @return mixed
      */
-    public function actionDelete($id)
-    {
-        $this->findModel($id)->delete();
-        return;
-       // return $this->redirect(['index']);
+    public function actionDelete($id,$num_factura,$id_proforma)
+    {        
+        $number_proforma_num = Proformaitem::findAll('proforma_num'==$num_factura);
+        $num_proforma_lines = count($number_proforma_num);
+        
+        if ($num_proforma_lines >= 1){
+             $this->findModel($id)->delete();
+                //return $this->redirect(['index']);
+          
+        }else{
+              Yii::$app->runAction('factura/delete',['id'=>$id_proforma]);
+        }
     }
 
     /**
